+++ app/routes/chats/route.tsx
@@ -1,0 +1,9 @@
+import { Outlet } from "react-router";
+
+export default function Chat() {
+  return (
+    <>
+      <Outlet />
+    </>
+  );
+}
